cleanup for scene click
